Revert "Hotfix/app/components"
