--- client/constants/data.js
@@ -1,79 +1,0 @@
-export const data = [
-  {
-    id: '1',
-    name: 'Adobong Manok',
-    username: '@juanluna',
-    ratings: 65,
-    image:
-      'https://yummyfood.ph/wp-content/uploads/2021/08/Chicken-Adobo-Recipe.jpg',
-    description: 'This is a description for this meal plan.',
-  },
-  {
-    id: '2',
-    name: 'Sinigang na Baboy',
-    username: '@semicolon',
-    ratings: 231,
-    image:
-      'https://www.kawalingpinoy.com/wp-content/uploads/2013/01/sinigang-baboy-7.jpg',
-    description: 'This is a description for this meal plan.',
-  },
-  {
-    id: '3',
-    name: 'Tortang Isda',
-    username: '@juandelacruz',
-    ratings: 12,
-    image:
-      'https://i0.wp.com/www.angsarap.net/wp-content/uploads/2020/09/Tortang-Dulong-Wide.jpg?fit=1080%2C720&ssl=1',
-    description: 'This is a description for this meal plan.',
-  },
-  {
-    id: '4',
-    name: 'Kare-kare',
-    username: '@xenonxenon',
-    ratings: 654,
-    image:
-      'https://api.lifegetsbetter.ph/uploads/recipes/featured/Kare-KAre.jpg',
-    description: 'This is a description for this meal plan.',
-  },
-  {
-    id: '5',
-    name: 'Lumpiang Shanghai',
-    username: '@semicolon',
-    ratings: 132,
-    image:
-      'https://images.aws.nestle.recipes/resized/8fa4b6a13f1d904c77b6f6da0cfe5c21_nks-b3-d02_640_944_531.jpg',
-    description: 'This is a description for this meal plan.',
-  },
-  {
-    id: '6',
-    name: 'Lumpiang Shanghai',
-    username: '@semicolon',
-    ratings: 132,
-    image:
-      'https://images.aws.nestle.recipes/resized/8fa4b6a13f1d904c77b6f6da0cfe5c21_nks-b3-d02_640_944_531.jpg',
-    description: 'This is a description for this meal plan.',
-  },
-];
-
-export const homeData = [
-  {
-    id: 1,
-    headerTitle: 'Recommended Recipes',
-    subTitle: 'Recommended',
-  },
-  {
-    id: 2,
-    headerTitle: 'Breakfast',
-    subTitle: 'Breakfast',
-  },
-  {
-    id: 3,
-    headerTitle: 'Lunch',
-    subTitle: 'Lunch',
-  },
-  {
-    id: 4,
-    headerTitle: 'Dinner',
-    subTitle: 'Dinner',
-  },
-];
