+++ client/constants/index.js
@@ -1,5 +1,4 @@
 import { COLORS, FONT, SIZES, SHADOWS } from './theme';
 import Logo from './images';
-import { data as DATA, homeData as HOMEDATA } from './data';
 
-export { DATA, HOMEDATA, Logo, COLORS, FONT, SIZES, SHADOWS };
+export { Logo, COLORS, FONT, SIZES, SHADOWS };
